# config/swagger.yaml
openapi: 3.0.1
info:
  title: npsstore API
  description: This is a Backend API of NBSStore
  version: 1.0.0
  termsOfService: http://quocbaoit.com/terms/
  contact:
    email: NBSstoreProject@quocbaoit.com
...
paths:
...
  /api/v1/comment: 
    post:
      summary:  User post new comment
      description: User post new comment
      tags:
        - Comment
      security:
        - bearerAuth: []
      requestBody:
        description: Comment payload
        required: true
        content:
          application/json:
            schema:
              type: object
              $ref: '#/components/schemas/CommentForm'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'  
  /api/v1/comment/{parentCommentId}/reply: 
    get:
      summary:  User, non-user get comment replies
      description: User, non-user get comment replies
      tags:
        - Comment
      parameters:
        - name: parentCommentId
          in: path
          description: Parent Comment Id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
  /api/v1/comment/{commentId}: 
    delete:
      summary:  User, admin delete comment
      description: User, admin delete comment
      tags:
        - Comment
      parameters:
        - name: commentId
          in: path
          description: Comment Id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
  /api/v1/comment/{commentId}/content: 
    patch:
      summary:  User, admin change comment
      description: User, admin change comment
      tags:
        - Comment
      parameters:
        - name: commentId
          in: path
          description: Comment Id
          required: true
          schema:
            type: string
      requestBody:
        description: Comment content
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  description: Comment content
                  nullable: false
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
  /api/v1/comment/{commentId}/disable: 
    patch:
      summary:  User, admin disable comment
      description: User, admin disable comment
      tags:
        - Comment
      parameters:
        - name: commentId
          in: path
          description: Comment Id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
  /api/v1/comment/{commentId}/restore: 
    patch:
      summary:  User, admin restore comment
      description: User, admin restore comment
      tags:
        - Comment
      parameters:
        - name: commentId
          in: path
          description: Comment Id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: error message
        stack:
          type: string
          description: error message description
...
    Timestamp:
      type: object
      properties:
        createAt:
          type: string
          description: The date and time when this record is created
          format: date-time
        updatedAt:
          type: string
          description: The date and time when this record is updated
          format: date-time
...
    Comment:
      type: object
      properties:
        _id: 
          type: string
          description: Cart Object id
        user: 
          type: string
          description: User id
        product:
          type: string 
          description: Product object id
        content: 
          type: string
          description: Comment content
        parentComment:
          type: string
          description: Parent comment object id
        replies:
          type: array
          items: 
            $ref: '#/components/schemas/Comment'
          description: Commnet replies
        isDisabled:
          type: boolean
          description: is Disabled or not
        timestamp:
          $ref: '#/components/schemas/Timestamp'
        __v:
          type: number
          description: Document version key
    CommentForm:
      type: object
      properties:
        productId:
          type: string 
          description: Product object id
        content: 
          type: string
          description: Comment content
        parentCommentId:
          type: string
          description: Parent comment object id
          nullable: true
  # 1) Define the security scheme type (HTTP bearer)
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: apiKey
      description: "<i>JWT Authorization header using the Bearer scheme.</i> <b>Example</b>: <i><u>Bearer 12345abcdef</u></i>"
      in: header          # can be "header", "query" or "cookie"
      name: Authorization # name of the header, query parameter or cookie
# 2) Apply the security globally to all operations
security:
  - bearerAuth: [] # use the same name as above